outsourced LearnAttack
